Restructured agent file and plotting_server
