+++ agents/ppo2.py
@@ -13,7 +13,6 @@ from datetime import datetime
 from custom_policies.policies import ThreeOf128NonShared,  OneShared55TwoValueOnePolicy
 
 import numpy as np
-import plotly.graph_objects as go # or plotly.express as px
 
 # Filter tensorflow version warnings
 import os
@@ -31,9 +30,10 @@ tf.get_logger().setLevel(logging.ERROR)
 
 
 class ppo2:
-    def __init__(self):
-        self.env = gym.make('DeepWellEnvSpher-v0')     #Options: 'DeepWellEnv-v0' 'DeepWellEnv-v2' 'DeepWellEnv3d-v0' 'DeepWellEnvSpher-v0'
-   
+    def __init__(self,env):
+        self.env = env     #Options: 'DeepWellEnv-v0' 'DeepWellEnv-v2' 'DeepWellEnv3d-v0' 'DeepWellEnvSpher-v0'
+
+
     #Get model either by training a new one or loading an old one
     def get_model(self):
         #sys.argv fetches arg when running "deepwellstart.ps1 -r arg". This is to make it possible to load,train or retrain the agent.
@@ -114,80 +114,6 @@ class ppo2:
         model.save(trained_models_path + model_name)                                                                                                   #Save the retrained model
         return model
 
-
-    def get_fig(self,model):
-        xcoord_list, ycoord_list, zcoord_list, info = self.get_path_from_model(model)
-        
-        fig = go.Figure(data=[go.Scatter3d(x=xcoord_list, y=ycoord_list, z=zcoord_list, mode='lines', name="Well path", line=dict(width=10.0))])
+    
 
 
-        x_targets = info['xtargets']
-        y_targets = info['ytargets']
-        z_targets = info['ztargets']
-        radius_targets = info['t_radius']
-
-        x_hazards = info['xhazards']
-        y_hazards = info['yhazards']
-        z_hazards = info['zhazards']
-        radius_hazards = info['h_radius']
-
-
-
-        for i in range(len(x_targets)):
-            self.plot_ball(fig, "Target", 'greens', x_targets[i], y_targets[i], z_targets[i], radius_targets[i])
-        
-        for i in range(len(x_hazards)):
-            self.plot_ball(fig, "Hazard", 'reds', x_hazards[i], y_hazards[i], z_hazards[i], radius_hazards[i])
-
-
-        fig.update_layout(
-            scene = dict(
-                xaxis = dict(nticks=4, range=[self.env.xmin,self.env.xmax],title_text="East",),
-                yaxis = dict(nticks=4, range=[self.env.ymin,self.env.ymax],title_text="North",),
-                zaxis = dict(nticks=4, range=[self.env.zmax,self.env.zmin],title_text="TVD",),
-            ),
-        )
-
-        print("Minimum total distance: ", info['min_dist'])
-        print("Distance traveled: ", info['tot_dist'])    
-        print("Target hits:     ", info['hits'])
-        return fig
-
-
-
-    #Test the trained model, run until done, return list of visited coords
-    def get_path_from_model(self, model):
-        obs = self.env.reset()
-        xcoord_list = [self.env.x]      #Initialize list of path coordinates with initial position
-        ycoord_list = [self.env.y]
-        zcoord_list = [self.env.z]
-        info = {}
-
-        while True:
-            action, _states = model.predict(obs)
-            obs, rewards, done, info = self.env.step(action)
-            
-            print("reward: ",rewards) 
-            xcoord_list.append(info['x'])
-            ycoord_list.append(info['y'])
-            zcoord_list.append(info['z'])
-            
-            if done: break
-
-        return xcoord_list, ycoord_list, zcoord_list, info
-  
-
-
-    def plot_ball(self, figure, name, color, x0, y0, z0, radius):
-            # Make data
-            u = np.linspace(0, 2 * np.pi, 100)
-            v = np.linspace(0, np.pi, 100)
-
-            x = x0 + radius * np.outer(np.cos(u), np.sin(v))
-            y = y0 + radius * np.outer(np.sin(u), np.sin(v))
-            z = z0 + radius * np.outer(np.ones(np.size(u)), np.cos(v))
-
-            # Plot the surface
-            figure.add_trace(
-                go.Surface(x=x, y=y, z=z, colorscale=color, showscale=False, name=name),)        
-
